Move lemmatizer algorithm changes back into RussianLemmatizer
