+++ spacy/lang/uk/lemmatizer.py
@@ -2,12 +2,12 @@ from typing import Optional, Callable
 
 from thinc.api import Model
 
+from ..ru.lemmatizer import RussianLemmatizer
 from ...pipeline.lemmatizer import lemmatizer_score
-from ...pipeline.pymorphy_lemmatizer import PyMorhpyLemmatizer
 from ...vocab import Vocab
 
 
-class UkrainianLemmatizer(PyMorhpyLemmatizer):
+class UkrainianLemmatizer(RussianLemmatizer):
     def __init__(
         self,
         vocab: Vocab,
@@ -18,6 +18,28 @@ class UkrainianLemmatizer(PyMorhpyLemmatizer):
         overwrite: bool = False,
         scorer: Optional[Callable] = lemmatizer_score,
     ) -> None:
+        if mode in {"pymorphy2", "pymorphy2_lookup"}:
+            try:
+                from pymorphy2 import MorphAnalyzer
+            except ImportError:
+                raise ImportError(
+                    "The Ukrainian lemmatizer mode 'pymorphy2' requires the "
+                    "pymorphy2 library and dictionaries. Install them with: "
+                    "pip install pymorphy2 pymorphy2-dicts-uk"
+                ) from None
+            if getattr(self, "_morph", None) is None:
+                self._morph = MorphAnalyzer(lang="uk")
+        elif mode == "pymorphy3":
+            try:
+                from pymorphy3 import MorphAnalyzer
+            except ImportError:
+                raise ImportError(
+                    "The Ukrainian lemmatizer mode 'pymorphy3' requires the "
+                    "pymorphy3 library and dictionaries. Install them with: "
+                    "pip install pymorphy3 pymorphy3-dicts-uk"
+                ) from None
+            if getattr(self, "_morph", None) is None:
+                self._morph = MorphAnalyzer(lang="uk")
         super().__init__(
-            vocab, model, "uk", name=name, mode=mode, overwrite=overwrite, scorer=scorer, 
+            vocab, model, name, mode=mode, overwrite=overwrite, scorer=scorer
         )
